+++ src/main/java/com/tickitz/backend/order/dto/UpdateOrderRequestDto.java
@@ -3,10 +3,12 @@ package com.tickitz.backend.order.dto;
 import com.tickitz.backend.order.entity.Order;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UpdateOrderRequestDto {
   private Long id;
-  private Long totalPrice;
+  private BigDecimal totalPrice;
   private Integer totalTicket;
   private Long usedPoint;
